+++ src/Parallel.php
@@ -184,10 +184,11 @@ class Parallel
      */
     public function execute(InputInterface $input, OutputInterface $output): void
     {
+        $this->output->setOutput($output);
         try {
             $this->initializeTaskStack($input->getOption('subnet'));
         } catch (TaskStackFactoryException $e) {
-            $this->output->errorMessage($output, $e->getMessage());
+            $this->output->errorMessage($e->getMessage());
             return;
         }
 
@@ -195,7 +196,7 @@ class Parallel
         $this->globalTaskLogger->prepareGlobal(count($this->taskStack->getStackedTasks()), $input->getOption('subnet'));
 
         $start = microtime(true);
-        $this->output->startMessage($output);
+        $this->output->startMessage();
 
         // Add all tasks to tasks data array
         foreach ($this->taskStack->getStackedTasks() as $stashedTask) {
@@ -223,7 +224,7 @@ class Parallel
                     }
 
                     // Redraw output when task finished
-                    $this->output->printToOutput($output, $this->data, microtime(true) - $start);
+                    $this->output->printToOutput($this->data, microtime(true) - $start);
                     $this->logDoneStackedTask($doneStackedTask);
                 }
             }
@@ -259,7 +260,7 @@ class Parallel
                             $this->logger->error($stackedTask->getTask()->getSanitizedName() . ': ' . StringHelper::sanitize($errorLine));
                         }
 
-                        $this->output->printToOutput($output, $this->data, microtime(true) - $start);
+                        $this->output->printToOutput($this->data, microtime(true) - $start);
                         return;
                     }
 
@@ -280,15 +281,15 @@ class Parallel
                     }
 
                     $this->buildTaskData($stackedTask, $data);
-                    $this->output->printToOutput($output, $this->data, microtime(true) - $start);
+                    $this->output->printToOutput($this->data, microtime(true) - $start);
                 });
                 usleep(floor($this->sleep * self::MICROSECONDS_IN_SECOND));
             }
         }
 
         $this->globalTaskLogger->processGlobal();
-        $this->output->printToOutput($output, $this->data, microtime(true) - $start);
-        $this->output->finishMessage($output, $this->data, microtime(true) - $start);
+        $this->output->printToOutput($this->data, microtime(true) - $start);
+        $this->output->finishMessage($this->data, microtime(true) - $start);
     }
 
     /**