+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Count of visible rows in output table
 
 ### Fixed
+- Flickering of table output significantly reduced!
+- Progress bar shows fractional progress
+- Summary table on finish instead of done tasks
 - PHP Deprecated: Implicit conversion from float to int loses precision
 
 ## 2.2.0 - 2021-11-02